hamming code generator runs
